- Added response for faulty files

+++ pyenzyme/restful/read.py
@@ -1,7 +1,7 @@
 # @Author: Jan Range
 # @Date:   2021-03-18 22:33:21
 # @Last Modified by:   Jan Range
-# @Last Modified time: 2021-03-24 14:26:14
+# @Last Modified time: 2021-04-07 09:39:27
 
 from flask import Flask, request, send_file, jsonify, Response
 from flask_restful import Resource, Api
@@ -30,23 +30,36 @@ class Read(MethodResource):
         Reads JSON formatted data and converts to an EnzymeML container.
         """
         
+        # check if the post request has the file part
+        if 'omex' not in request.files:
+            return jsonify( {"response": 'No file part'} )
+        
         # receive OMEX file
-        file = request.files['omex'].read()
+        file = request.files['omex']
         
-        # Write to temp file
-        dirpath = os.path.join( os.path.dirname( os.path.realpath(__file__)), "read_temp" )
-        os.makedirs(dirpath, exist_ok=True)
+        # if user does not select file, browser also
+        # submit an empty part without filename
+        if file.filename == '':
+            return jsonify( {"response": 'No file selected'} )
         
-        tmp = os.path.join( dirpath, next(tempfile._get_candidate_names()) )
+        if file and file.filename.split('.')[-1] == "omex":
+            
+            file = file.read()
+            
+            # Write to temp file
+            dirpath = os.path.join( os.path.dirname( os.path.realpath(__file__)), "read_temp" )
+            os.makedirs(dirpath, exist_ok=True)
+            
+            tmp = os.path.join( dirpath, next(tempfile._get_candidate_names()) )
 
-        with open(tmp, 'wb') as f:
-            f.write(file)
-        
-        # Save JSON in variable
-        enzmldoc = EnzymeMLReader().readFromFile(tmp)
-        JSON = enzmldoc.toJSON(d=True)
-        
-        # remove temp file
-        os.remove(tmp)
-        
-        return Response(json.dumps(JSON),  mimetype='application/json')
+            with open(tmp, 'wb') as f:
+                f.write(file)
+            
+            # Save JSON in variable
+            enzmldoc = EnzymeMLReader().readFromFile(tmp)
+            JSON = enzmldoc.toJSON(d=True)
+            
+            # remove temp file
+            os.remove(tmp)
+            
+            return Response(json.dumps(JSON),  mimetype='application/json')
